Refactor project and build profeat ext_modules
