+++ annopro/data_procession/_libprofeatconfig/input-pli.dat
@@ -1,0 +1,3 @@
+SYC1_MYCTU         +    1728-95-6                                             
+FABPL_PIG    +    91-08-7
+ARGR_ECOLI         +    146795-42-8
