+++ annopro/data_procession/_libprofeatconfig/input-ppi.dat
@@ -1,0 +1,3 @@
+SYC1_MYCTU  +  ARGR_ECOLI
+SYC1_MYCTU  +  FABPL_PIG
+ARGR_ECOLI +  FABPL_PIG
