+++ setup.cfg
@@ -1,0 +1,12 @@
+# Inside of setup.cfg
+[metadata]
+name = annopro
+version = 0.0.1
+description = A simple python package for annotating protein sequences
+long_description = file: README.md
+long_description_content_type = text/markdown
+classifiers = 
+    Programming Language :: Python :: 3
+    License :: OSI Approved :: MIT License
+    Topic :: Scientific/Engineering :: Bio-Informatics
+license = MIT License
